// src/generated/types.gen.ts
export type Resource = {
    /**
     * Type of the supported resources
     */
    type: string;
    /**
     * Unique identifier representing a specific resource instance
     */
    id: string;
    /**
     * Clip v1 resource identifier
     */
    id_v1?: string;
};

export type ResourceIdentifier = {
    /**
     * The unique id of the referenced resource
     */
    rid: string;
    /**
     * The type of the referenced resource
     */
    rtype: 'device' | 'bridge_home' | 'room' | 'zone' | 'light' | 'button' | 'relative_rotary' | 'temperature' | 'light_level' | 'motion' | 'camera_motion' | 'entertainment' | 'contact' | 'tamper' | 'grouped_light' | 'device_power' | 'zigbee_bridge_connectivity' | 'zigbee_connectivity' | 'zgp_connectivity' | 'bridge' | 'zigbee_device_discovery' | 'homekit' | 'matter' | 'matter_fabric' | 'scene' | 'entertainment_configuration' | 'public_image' | 'auth_v1' | 'behavior_script' | 'behavior_instance' | 'geofence' | 'geofence_client' | 'geolocation' | 'smart_scene' | 'grouped_motion' | 'grouped_light_level';
};

/**
 * Common resource properties including the owner
 */
export type ResourceOwned = Resource & {
    owner: ResourceIdentifier;
};
// ...
export type ProductArchetype = 'bridge_v2' | 'unknown_archetype' | 'classic_bulb' | 'sultan_bulb' | 'flood_bulb' | 'spot_bulb' | 'candle_bulb' | 'luster_bulb' | 'pendant_round' | 'pendant_long' | 'ceiling_round' | 'ceiling_square' | 'floor_shade' | 'floor_lantern' | 'table_shade' | 'recessed_ceiling' | 'recessed_floor' | 'single_spot' | 'double_spot' | 'table_wash' | 'wall_lantern' | 'wall_shade' | 'flexible_lamp' | 'ground_spot' | 'wall_spot' | 'plug' | 'hue_go' | 'hue_lightstrip' | 'hue_iris' | 'hue_bloom' | 'bollard' | 'wall_washer' | 'hue_play' | 'vintage_bulb' | 'vintage_candle_bulb' | 'ellipse_bulb' | 'triangle_bulb' | 'small_globe_bulb' | 'large_globe_bulb' | 'edison_bulb' | 'christmas_tree' | 'string_light' | 'hue_centris' | 'hue_lightstrip_tv' | 'hue_lightstrip_pc' | 'hue_tube' | 'hue_signe' | 'pendant_spot' | 'ceiling_horizontal' | 'ceiling_tube';

export type ProductData = {
    /**
     * Unique identification of device model
     */
    model_id: string;
    /**
     * Name of device manufacturer
     */
    manufacturer_name: string;
    /**
     * Name of the product
     */
    product_name: string;
    product_archetype: ProductArchetype;
    /**
     * This device is Hue certified
     */
    certified: boolean;
    /**
     * Software version of the product
     */
    software_version: string;
    /**
     * Hardware type; identified by Manufacturer code and ImageType
     */
    hardware_platform_type?: string;
};

/**
 * Definition of a device resource
 */
export type DeviceGet = ResourceOwned & {
    type?: 'device';
    product_data: ProductData;
    metadata: {
        /**
         * Human readable name of a resource
         */
        name: string;
        archetype: ProductArchetype;
    };
    /**
     * Triggers a visual identification sequence on the device
     */
    identify: {
        [key: string]: unknown;
    };
    usertest?: {
        status?: 'set' | 'changing';
        /**
         * Activates or extends user usertest mode of device for 120 seconds.
         * `false` deactivates usertest mode.
         * In usertest mode, devices report changes in state faster and indicate state changes on device LED (if applicable)
         *
         */
        usertest?: boolean;
    };
    /**
     * References all services providing control and state of the device.
     */
    services: Array<ResourceIdentifier>;
};
// ...
export type RoomArchetype = 'living_room' | 'kitchen' | 'dining' | 'bedroom' | 'kids_bedroom' | 'bathroom' | 'nursery' | 'recreation' | 'office' | 'gym' | 'hallway' | 'toilet' | 'front_door' | 'garage' | 'terrace' | 'garden' | 'driveway' | 'carport' | 'home' | 'downstairs' | 'upstairs' | 'top_floor' | 'attic' | 'guest_room' | 'staircase' | 'lounge' | 'man_cave' | 'computer' | 'studio' | 'music' | 'tv' | 'reading' | 'closet' | 'storage' | 'laundry_room' | 'balcony' | 'porch' | 'barbecue' | 'pool' | 'other';

export type RoomGet = Resource & {
    /**
     * Child devices/services to group by the derived group
     */
    children: Array<ResourceIdentifier>;
    /**
     * References all services aggregating control and state of children in the group.
     * This includes all services grouped in the group hierarchy given by child relation.
     * This includes all services of a device grouped in the group hierarchy given by child relation.
     * Aggregation is per service type, ie every service type which can be grouped has a corresponding definition of
     * grouped type.
     * Supported types: – grouped_light
     *
     */
    services: Array<ResourceIdentifier>;
    /**
     * configuration object for a room
     */
    metadata: {
        /**
         * Human readable name of a resource
         */
        name: string;
        archetype: RoomArchetype;
    };
};